pyproject.toml
[project]
name = "app"
version = "0.0.1"
description = ""
authors = [
    {name = "LJ"}
]
license = {text = "None"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "watchdog (>=6.0.0,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pynput (>=1.8.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
